Close popper when click/touch starts
Previously, when starting a click inside the popper and dragging a
slider (e.g., color-picker) to the edge and outside of the popper, the
click events end position would be the target, which would
unintendedly cause the popper to close.

For consistency in the popper's closing behavior, this is set as the
default for all input-popper components, not just those affected by
this issue.