Allow extending slots from Admin components
Extending a Comet Admin component using `styled()` requires the
`className` prop to be passed through to the components `root` slot.

Also added the `RootSlotProps` type to use as a base for components props.
This abstracts away the need to manually define the `sx`, `className`
and `slotProps` props in every component.